+++ lib/presentation/screens/main_screen.dart
@@ -4,8 +4,8 @@ import 'package:betrade/core/theme/app_text_style.dart';
 import 'package:betrade/presentation/screens/explore/explore_page.dart';
 import 'package:betrade/presentation/screens/homeScreen/HomeScreen.dart';
 import 'package:betrade/presentation/screens/portfolio/portfolio_page.dart';
-import 'package:betrade/presentation/screens/profile/info_chart_screen.dart';
 import 'package:betrade/presentation/screens/profile/profile_page.dart';
+import 'package:betrade/presentation/screens/rankings/rankings_screen.dart';
 import 'package:betrade/presentation/screens/verification/verify_account.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -320,7 +320,7 @@ class _MainScreenState extends State<MainScreen> {
             onBannerTap: _showWelcomePopup,
           ),
           ExplorePage(),
-          InfoChartScreen(),
+          RankingsScreen(),
           PortfolioPage(),
           ProfilePage(),
         ],